decimal -> userassessment

+++ SARMS/Users/Account.cs
@@ -260,7 +260,7 @@ namespace SARMS.Users
                         if (reader.HasRows)
                         {
                             reader.Read();
-                            result.Add(new StudentAssessment() { account = student, Assessment = ass, Mark = Convert.ToDecimal(reader[0]) });
+                            result.Add(new StudentAssessment() { account = student, Assessment = ass, Mark = Convert.ToDouble(reader[0]) });
                         }
                         reader.Close();
                     }
